app_aggregator/requirements.txt
pydantic
aerospike
kafka-python
python-snappy
crc32c
jsonpickle
python-dateutil
